Improved Card Manager and Fixed Bugs

+++ CardManager.java
@@ -1,33 +1,48 @@
 package group_project;
 
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 
 public class CardManager 
 {
+    //Can change, but manager doesn't manage sizing since it is easier to do within the table class
     private DblTableAutoResize<List<Integer>, AttributeSet> table;
     private MapManager<CardObject> manager;    
 
-    public CardManager(CardObject... cards)
+    //Creates a card manager with a 2 power doubling table
+    public CardManager()
     {
         table = new DblTableAutoResize<>(2, 0.75);
         manager = new MapManager<>(table);
     }
 
-    public int addCard(CardObject cd)
+    //Adds card to table. I'm lazy so it uses set if keepid; add and sets to newId otherwise
+    public int addCard(CardObject cd, boolean keepId)
     {
         AttributeSet temp = new AttributeSet(cd.getAtributes());
+        if(keepId)
+        {
+            manager.set(cd.getId(),temp,cd);
+            return cd.getId();
+        }
         int newId = manager.putObject(cd, temp);
         cd.setId(newId);
         return newId;
     }
 
+    //Removes card with id provided
     public void removeCard(CardObject cd)
     {
         manager.removeId(cd.getId(),new AttributeSet(cd.getAtributes()));
     }
 
+    //Grabs all Card Objects matching attributes provided
     public List<CardObject> getMatching(AttributeSet attributes)
     {
+        //If attributes is null, get all cards
+        if(attributes==null)
+            attributes = new AttributeSet(new ArrayList<>());
         return manager.get(attributes);
     }
 
@@ -36,6 +51,28 @@ public class CardManager
         return manager.numObjects();
     }
 
+    //Adds multiple cards to db
+    public void addMultiple(boolean keepId, CardObject... cards)
+    {
+        for(CardObject c : cards)
+        {
+            addCard(c,keepId);
+        }
+    }
+
+    //Returns map manager toString
+    public String toString()
+    {
+        String temp = manager.toString();
+        return temp;
+    }
+
+    public void addMultiple(boolean keepId, Collection<CardObject> cards)
+    {
+        addMultiple(keepId, (CardObject[])cards.toArray());
+    }
+
+    //Just a lazy way to manage table sizing
     private class DblTableAutoResize<T,U> extends DblHashMap_HashCode<T, U>
     {
         private int size;
@@ -47,6 +84,7 @@ public class CardManager
             this.loadFactor = loadFactor;
         }
 
+        //Just doubles table once beyond load factor
         @Override
         public void put(T value, U key) {
             super.put(value, key);
@@ -55,6 +93,5 @@ public class CardManager
             size = size*2;
             resize(size);
         }
-        
     }
 }
